+++ src/KingmakerGunslinger/Blueprints/GunslingerClassBlueprints.cs
@@ -26,7 +26,8 @@ namespace KingmakerGunslinger.Blueprints
             StartlingShotBlueprintSet startlingShot,
             TargetingHeadBlueprintSet targetingHead,
             TargetingTorsoBlueprintSet targetingTorso,
-            TargetingLegsBlueprintSet targetingLegs)
+            TargetingLegsBlueprintSet targetingLegs,
+            BleedingWoundBlueprintSet bleedingWound)
         {
             CharacterClass = characterClass ?? throw new ArgumentNullException("characterClass");
             Progression = progression ?? throw new ArgumentNullException("progression");
@@ -49,6 +50,8 @@ namespace KingmakerGunslinger.Blueprints
                 throw new ArgumentNullException("targetingTorso");
             TargetingLegs = targetingLegs ??
                 throw new ArgumentNullException("targetingLegs");
+            BleedingWound = bleedingWound ??
+                throw new ArgumentNullException("bleedingWound");
         }
         internal BlueprintCharacterClass CharacterClass { get; private set; }
         internal BlueprintProgression Progression { get; private set; }
@@ -67,7 +70,8 @@ namespace KingmakerGunslinger.Blueprints
         internal TargetingHeadBlueprintSet TargetingHead { get; private set; }
         internal TargetingTorsoBlueprintSet TargetingTorso { get; private set; }
         internal TargetingLegsBlueprintSet TargetingLegs { get; private set; }
-        internal int Count { get { return 4 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count + TargetingLegs.Count; } }
+        internal BleedingWoundBlueprintSet BleedingWound { get; private set; }
+        internal int Count { get { return 4 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count + TargetingLegs.Count + BleedingWound.Count; } }
     }
 
     internal sealed class GunslingerClassCatalogPublication
@@ -167,6 +171,8 @@ namespace KingmakerGunslinger.Blueprints
                 TargetingTorsoBlueprints.Register(registry);
             TargetingLegsBlueprintSet targetingLegs =
                 TargetingLegsBlueprints.Register(registry);
+            BleedingWoundBlueprintSet bleedingWound =
+                BleedingWoundBlueprints.Register(registry);
             BlueprintProgression progression = registry.Register<BlueprintProgression>(
                 ProgressionSymbol, () => CreateProgression());
 
@@ -186,13 +192,14 @@ namespace KingmakerGunslinger.Blueprints
             progression.LevelEntries[6].Features.Add(targetingHead.Feature);
             progression.LevelEntries[6].Features.Add(targetingTorso.Feature);
             progression.LevelEntries[6].Features.Add(targetingLegs.Feature);
+            progression.LevelEntries[10].Features.Add(bleedingWound.Feature);
             Validate(characterClass, progression, proficiencies, fullBab, goodSave,
                 poorSave, startingPistol, blackPowder, leadBall,
                 simple, martial, lightArmor, firearmProficiency);
             return new GunslingerClassBlueprintSet(characterClass, progression,
                 proficiencies, grit, deadeye, dodge, quickClear, nimble, initiative,
                 pistolWhip, utilityShot, gunTraining, deadShot, startlingShot,
-                targetingHead, targetingTorso, targetingLegs);
+                targetingHead, targetingTorso, targetingLegs, bleedingWound);
         }
 
         internal static GunslingerClassCatalogPublication Publish(